refactor: Make Reporting event message flags reusable
