+++ src/main/k8s/reporting_v2.cue
@@ -309,7 +309,7 @@ package k8s
 						_metricSpecConfigFileFlag,
 						"--port=8443",
 						"--health-port=8080",
-			] + _tlsArgs + _internalApiTarget.args + _kingdomApiTarget.args + _accessApiTarget.args
+			] + _tlsArgs + _internalApiTarget.args + _kingdomApiTarget.args + _accessApiTarget.args + _eventDescriptorArgs
 			spec: {
 				jobTemplate: spec: template: spec: _mounts: {
 					"mc-config": {
