+++ src/main/kotlin/org/wfanet/measurement/reporting/deploy/v2/common/BUILD.bazel
@@ -68,3 +68,16 @@ kt_jvm_library(
         "@wfa_common_jvm//src/main/kotlin/org/wfanet/measurement/common/grpc",
     ],
 )
+
+kt_jvm_library(
+    name = "event_message_flags",
+    srcs = ["EventMessageFlags.kt"],
+    deps = [
+        "//src/main/kotlin/org/wfanet/measurement/reporting/service/api/v2alpha:event_descriptor",
+        "//src/main/proto/wfa/measurement/api/v2alpha:event_annotations_kt_jvm_proto",
+        "//src/main/proto/wfa/measurement/api/v2alpha:media_type_kt_jvm_proto",
+        "@wfa_common_jvm//imports/java/com/google/protobuf",
+        "@wfa_common_jvm//imports/java/picocli",
+        "@wfa_common_jvm//src/main/kotlin/org/wfanet/measurement/common",
+    ],
+)
