+++ src/main/kotlin/org/wfanet/measurement/reporting/deploy/v2/common/server/V2AlphaPublicApiServer.kt
@@ -18,8 +18,6 @@ package org.wfanet.measurement.reporting.deploy.v2.common.server
 
 import com.google.protobuf.DescriptorProtos
 import com.google.protobuf.Descriptors
-import com.google.protobuf.ExtensionRegistry
-import com.google.protobuf.TypeRegistry
 import com.google.protobuf.util.JsonFormat
 import io.grpc.Channel
 import io.grpc.ServerServiceDefinition
@@ -41,13 +39,11 @@ import org.wfanet.measurement.access.v1alpha.PermissionsGrpcKt
 import org.wfanet.measurement.access.v1alpha.PrincipalsGrpcKt
 import org.wfanet.measurement.api.v2alpha.CertificatesGrpcKt.CertificatesCoroutineStub as KingdomCertificatesCoroutineStub
 import org.wfanet.measurement.api.v2alpha.DataProvidersGrpcKt.DataProvidersCoroutineStub as KingdomDataProvidersCoroutineStub
-import org.wfanet.measurement.api.v2alpha.EventAnnotationsProto
 import org.wfanet.measurement.api.v2alpha.EventGroupMetadataDescriptorsGrpcKt.EventGroupMetadataDescriptorsCoroutineStub as KingdomEventGroupMetadataDescriptorsCoroutineStub
 import org.wfanet.measurement.api.v2alpha.EventGroupsGrpcKt.EventGroupsCoroutineStub as KingdomEventGroupsCoroutineStub
 import org.wfanet.measurement.api.v2alpha.MeasurementConsumerKey
 import org.wfanet.measurement.api.v2alpha.MeasurementConsumersGrpcKt.MeasurementConsumersCoroutineStub as KingdomMeasurementConsumersCoroutineStub
 import org.wfanet.measurement.api.v2alpha.MeasurementsGrpcKt.MeasurementsCoroutineStub as KingdomMeasurementsCoroutineStub
-import org.wfanet.measurement.api.v2alpha.MediaTypeProto
 import org.wfanet.measurement.api.v2alpha.ModelLinesGrpcKt.ModelLinesCoroutineStub as KingdomModelLinesCoroutineStub
 import org.wfanet.measurement.api.withAuthenticationKey
 import org.wfanet.measurement.common.ProtoReflection
@@ -77,6 +73,7 @@ import org.wfanet.measurement.internal.reporting.v2.ReportsGrpcKt.ReportsCorouti
 import org.wfanet.measurement.internal.reporting.v2.measurementConsumer
 import org.wfanet.measurement.measurementconsumer.stats.VariancesImpl
 import org.wfanet.measurement.reporting.deploy.v2.common.EncryptionKeyPairMap
+import org.wfanet.measurement.reporting.deploy.v2.common.EventMessageFlags
 import org.wfanet.measurement.reporting.deploy.v2.common.InProcessServersMethods.startInProcessServerWithService
 import org.wfanet.measurement.reporting.deploy.v2.common.KingdomApiFlags
 import org.wfanet.measurement.reporting.deploy.v2.common.ReportingApiServerFlags
@@ -85,7 +82,6 @@ import org.wfanet.measurement.reporting.service.api.CelEnvCacheProvider
 import org.wfanet.measurement.reporting.service.api.InMemoryEncryptionKeyPairStore
 import org.wfanet.measurement.reporting.service.api.v2alpha.BasicReportsService
 import org.wfanet.measurement.reporting.service.api.v2alpha.DataProvidersService
-import org.wfanet.measurement.reporting.service.api.v2alpha.EventDescriptor
 import org.wfanet.measurement.reporting.service.api.v2alpha.EventGroupMetadataDescriptorsService
 import org.wfanet.measurement.reporting.service.api.v2alpha.EventGroupsService
 import org.wfanet.measurement.reporting.service.api.v2alpha.MetricCalculationSpecsService
@@ -119,6 +115,7 @@ private object V2AlphaPublicApiServer {
     @CommandLine.Mixin v2AlphaFlags: V2AlphaFlags,
     @CommandLine.Mixin v2AlphaPublicServerFlags: V2AlphaPublicServerFlags,
     @CommandLine.Mixin encryptionKeyPairMap: EncryptionKeyPairMap,
+    @CommandLine.Mixin eventMessageFlags: EventMessageFlags,
     @CommandLine.Option(
       names = ["--system-measurement-consumer"],
       description =
@@ -301,25 +298,6 @@ private object V2AlphaPublicApiServer {
         .build()
         .withShutdownTimeout(Duration.ofSeconds(30))
 
-    val eventDescriptor: EventDescriptor? =
-      // TODO(@tristanvuong2021): Flags will be required once BasicReports Phase 2 is completed.
-      if (
-        v2AlphaPublicServerFlags.eventMessageTypeUrl.isNotEmpty() &&
-          v2AlphaPublicServerFlags.eventMessageDescriptorSetFiles.isNotEmpty()
-      ) {
-        val eventDescriptor: Descriptors.Descriptor =
-          checkNotNull(
-            buildTypeRegistry(v2AlphaPublicServerFlags.eventMessageDescriptorSetFiles)
-              .getDescriptorForTypeUrl(v2AlphaPublicServerFlags.eventMessageTypeUrl)
-          ) {
-            "--event-message-type-url is invalid"
-          }
-
-        EventDescriptor(eventDescriptor)
-      } else {
-        null
-      }
-
     val services: List<ServerServiceDefinition> =
       listOf(
         DataProvidersService(
@@ -385,7 +363,7 @@ private object V2AlphaPublicApiServer {
             InternalReportingSetsCoroutineStub(channel),
             InternalMetricCalculationSpecsCoroutineStub(channel),
             ReportsCoroutineStub(inProcessReportsChannel),
-            eventDescriptor,
+            eventMessageFlags.eventDescriptor,
             basicReportMetricSpecConfig,
             SecureRandom().asKotlinRandom(),
             authorization,
@@ -458,55 +436,7 @@ private object V2AlphaPublicApiServer {
 
     lateinit var knownEventGroupMetadataTypes: List<Descriptors.FileDescriptor>
       private set
-
-    @CommandLine.Option(
-      names = ["--event-message-type-url"],
-      description = ["Fully qualified name of the event message type."],
-      required = false,
-      defaultValue = "",
-    )
-    lateinit var eventMessageTypeUrl: String
-      private set
-
-    @CommandLine.Option(
-      names = ["--event-message-descriptor-set"],
-      description =
-        [
-          "Path to a serialized FileDescriptorSet containing an event message type and/or its " +
-            "dependencies.",
-          "This can be specified multiple times.",
-        ],
-      required = false,
-    )
-    var eventMessageDescriptorSetFiles: List<File> = emptyList()
-      private set
   }
-
-  private fun buildTypeRegistry(descriptorSetFiles: List<File>): TypeRegistry {
-    val descriptorSets: List<DescriptorProtos.FileDescriptorSet> =
-      descriptorSetFiles.map {
-        it.inputStream().use { input ->
-          DescriptorProtos.FileDescriptorSet.parseFrom(input, EXTENSION_REGISTRY)
-        }
-      }
-
-    return TypeRegistry.newBuilder()
-      .add(ProtoReflection.buildDescriptors(descriptorSets, KNOWN_TYPES))
-      .build()
-  }
-
-  private val KNOWN_TYPES =
-    ProtoReflection.WELL_KNOWN_TYPES +
-      EventAnnotationsProto.getDescriptor() +
-      MediaTypeProto.getDescriptor()
-
-  private val EXTENSION_REGISTRY =
-    ExtensionRegistry.newInstance()
-      .apply {
-        add(EventAnnotationsProto.eventTemplate)
-        add(EventAnnotationsProto.templateField)
-      }
-      .unmodifiable
 }
 
 fun main(args: Array<String>) = commandLineMain(V2AlphaPublicApiServer::run, args)
